Clean up package dependency references
All the `package.json` references to `tko.*` should go to the appropriate subdirectory of the tko monorepo.  In the interim they’re at least homogenized.

It turns out for karma and rollup lerna needs to be bootstrapped with `—hoist`.

Karma’s ComonJS plugin cannot be limited by the `include`; it appears that “node_modules/core-js/” does not match “node_modules/**”.  Go figure.

Signed-off-by: Brian M Hunt <bmh@tko.io>