LifeCycle) Fix the `mixInto` functionality for ES6
In ES6 (without polyfill/Babel/transpilation/etc), the `LifeCycle.prototype` properties (i.e. prototype properties of a Class) are not enumerable.

To work around this we can use `Object.getOwnProperty{Names/Descriptors}` to do what is basically an `Object.assign` of non-enumerable properties.

Turns out this issue of non-enumerable prototype ES6 class properties has been a source of some consternation and debate.  Most (all) of the polyfills I saw did not recognize this problem, and relied on `for-in`, which only works for enumerable properties - hence, not here.