tko.bind) Fix issue with the dependency chain and huge performance problem
This addresses the on-again/off-again performance problem experienced and noted here: https://stackoverflow.com/questions/44414554

The synchronous `applyBindings` that was being called in various bindings (e.g. `template`, `if`, `with`) was being repeated for all child bindings, causing large refreshes.

The problem was caused when those bindings (`template`, `if`, etc) were changed to be descendants of `BindingHandler`, and were no longer converted via the `LegacyBindingHandler`.  What was the `init` function of those bindings became the `constructor` method of the new handler, but with the important difference that the `constructor` was not being called inside an `dependencyDetection.ignore`.