binding.attr) Fix mis-using of `style` attribute
This comes from mirroring the ko.punches code here:

https://github.com/mbest/knockout.punches/blob/master/src/interpolationMarkup.js#L163-L169 — basically:

```
    if (ko.getBindingHandler(name)) {
        return name + ':' + value;
    } else {
        return 'attr.' + name + ':' + value;
    }
}
```

The way this code works, any binding handler with the same name as an DOM node attribute would - instead of interpolation - call the bindinghandler with the value of the interpolation - which is generally not the desired outcome.

A white-list was added for those attributes who properly map to a binding handler (i.e. `value` and `checked`, though certainly there could be others).