npm) Update to latest packages, add globals, use .es6 knockout for testing
If we use `dist/knockout.js` for testing the knockout package, then we have to recompile every single dependency.  If we attempt to link directly to the files `src/index.js` as we do for the .es6 variants then Typescript does not inject its tslib dependencies.

So we test off of `dist/knockout.es6.js`.  One still has to recompile knockout to re-test every change.