jsx/slots) Support slots with JSX
The problem with slots and JSX is that JSX stores the attributes for bindings on the nodes directly (via the `NATIVE_BINDINGS` Symbol in NativeProvider).  When the nodes from the slots are cloned into the template with `document.cloneNode`, those symbols are not cloned.

A `cloneNode` function in JSX remedies this for nodes that have a `slot` attribute. We only do this for `slot` nodes  to save memory, let we have a copy of the original JSX AST for every JSX node.